game classifier added
The results as follow:

For each season we get a top 3 for each team depending on overall minutes.

For each game we look at that team’s top 3 for that season. 

If they all play 10+ minutes, they are the top 3.

If anyone plays less than 10 minutes, the top 3 revert to the three players who play the most minutes. 

Majority of the data uses the actually season long top 3, so this will be useful to train on